+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Assets are not served under the /assets directory
+- Return a JSON response from all API calls (except assets)
 
 ## [v0.10.0](https://github.com/elwinar/rcoredump/releases/tag/v0.10.0) - 2020-04-25
 ### Added