+++ CHANGELOG.md
@@ -5,12 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [v0.11.0](https://github.com/elwinar/rcoredump/releases/tag/v0.11.0) - 2020-05-04
 ### Added
-- Add a delay parameter to every API endpoint to allow testing under slow conditions
-- Add a special message on the app first load to avoid confusion with empty results
+- Delay parameter to every API endpoint to allow testing under slow conditions
+- Special message on the app first load to avoid confusion with empty results
 - "Delete Core" button in the webapp
 ### Changed
-- Assets are not served under the /assets directory
+- Assets are served under the /assets directory
 - Return a JSON response from all API calls (except assets)
 
 ## [v0.10.0](https://github.com/elwinar/rcoredump/releases/tag/v0.10.0) - 2020-04-25